Move CompoundException to the Fixie.Internal namespace, now that it is only used internally and no longer exposed publicly in test results.
