+++ src/Fixie/Fixie.csproj
@@ -62,7 +62,7 @@
     <Compile Include="Internal\ConstructionFrequency.cs" />
     <Compile Include="Options.cs" />
     <Compile Include="ParameterSource.cs" />
-    <Compile Include="Execution\CompoundException.cs" />
+    <Compile Include="Internal\CompoundException.cs" />
     <Compile Include="TestAssembly.cs" />
     <Compile Include="Conventions\ClassExpression.cs" />
     <Compile Include="Internal\MethodDiscoverer.cs" />
